Moved providers lib folder, added empty spec files
