+++ lib/omnisharp-atom/features/go-to-definition.ts
@@ -34,7 +34,7 @@ class GoToDefinition implements OmniSharp.IFeature {
                 Observable.fromEventPattern(x => { (<any>atom.getCurrentWindow()).on('blur', x) }, x => { (<any>atom.getCurrentWindow()).removeListener('blur', x) }),
                 Observable.fromEvent<KeyboardEvent>(view[0], 'keyup')
                     .where(x => x.which === 17 || x.which === 224 || x.which === 93 || x.which === 91)
-                )
+            )
                 .throttleFirst(100);
 
             var keydown = Observable.fromEvent<KeyboardEvent>(view[0], 'keydown')
@@ -113,8 +113,7 @@ class GoToDefinition implements OmniSharp.IFeature {
                             searchAllPanes: true
                         });
                     } else {
-                        atom.emitter.emit("omnisharp-atom:error",
-                            "Can't navigate to '" + word + "'");
+                        atom.notifications.addWarning("Can't navigate to '" + word + "'");
                     }
                 });
         }
