+++ lib/omnisharp-atom/providers/apply-changes.ts
