+++ lib/omnisharp-atom/providers/omni-select-list-view.ts
