+++ tsconfig.json
@@ -48,18 +48,11 @@
         "./lib/omnisharp-atom/features/code-check.ts",
         "./lib/omnisharp-atom/features/code-format.ts",
         "./lib/omnisharp-atom/features/command-runner.ts",
-        "./lib/omnisharp-atom/features/error-handler.ts",
         "./lib/omnisharp-atom/features/find-symbols.ts",
         "./lib/omnisharp-atom/features/find-usages.ts",
         "./lib/omnisharp-atom/features/go-to-definition.ts",
         "./lib/omnisharp-atom/features/highlight.ts",
         "./lib/omnisharp-atom/features/intellisense.ts",
-        "./lib/omnisharp-atom/features/lib/apply-changes.ts",
-        "./lib/omnisharp-atom/features/lib/completion-provider.ts",
-        "./lib/omnisharp-atom/features/lib/framework-provider.ts",
-        "./lib/omnisharp-atom/features/lib/linter-provider.ts",
-        "./lib/omnisharp-atom/features/lib/omni-select-list-view.ts",
-        "./lib/omnisharp-atom/features/lib/project-provider.ts",
         "./lib/omnisharp-atom/features/log.ts",
         "./lib/omnisharp-atom/features/lookup.ts",
         "./lib/omnisharp-atom/features/navigate-up-down.ts",
@@ -71,6 +64,12 @@
         "./lib/omnisharp-atom/features/server-information.ts",
         "./lib/omnisharp-atom/features/solution-information.ts",
         "./lib/omnisharp-atom/omnisharp-atom.ts",
+        "./lib/omnisharp-atom/providers/apply-changes.ts",
+        "./lib/omnisharp-atom/providers/completion-provider.ts",
+        "./lib/omnisharp-atom/providers/framework-provider.ts",
+        "./lib/omnisharp-atom/providers/linter-provider.ts",
+        "./lib/omnisharp-atom/providers/omni-select-list-view.ts",
+        "./lib/omnisharp-atom/providers/project-provider.ts",
         "./lib/omnisharp-atom/views/build-output-pane-view.ts",
         "./lib/omnisharp-atom/views/code-actions-view.ts",
         "./lib/omnisharp-atom/views/codecheck-output-pane-view.ts",
@@ -88,8 +87,26 @@
         "./lib/omnisharp-atom/views/tooltip-view.ts",
         "./lib/omnisharp-atom/world.ts",
         "./lib/typings.d.ts",
+        "./spec/atom/dock-spec.ts",
+        "./spec/atom/framework-selector-spec.ts",
+        "./spec/atom/status-bar-spec.ts",
+        "./spec/atom/update-project-spec.ts",
         "./spec/features/code-format-spec.ts",
         "./spec/features/command-runner-spec.ts",
+        "./spec/features/find-symbols-spec.ts",
+        "./spec/features/find-usages-spec.ts",
+        "./spec/features/go-to-definition-spec.ts",
+        "./spec/features/highlight-spec.ts",
+        "./spec/features/intellisense-spec.ts",
+        "./spec/features/log-spec.ts",
+        "./spec/features/lookup-spec.ts",
+        "./spec/features/navigate-up-down-spec.ts",
+        "./spec/features/package-restore-spec.ts",
+        "./spec/features/project-lock-spec.ts",
+        "./spec/features/rename-spec.ts",
+        "./spec/features/run-tests-spec.ts",
+        "./spec/features/server-information-spec.ts",
+        "./spec/features/solution-information-spec.ts",
         "./spec/omnisharp-atom-spec.ts",
         "./spec/test-helpers.ts",
         "./spec/tsd.d.ts",
